# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/sections/{sectionId}/answers": {
      "post": {
        "operationId": "AnswersController_create_v1",
        "parameters": [
          {
            "name": "sectionId",
            "required": true,
            "in": "path",
            "description": "Section UUID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnswerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Answer created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not owner or admin"
          },
          "404": {
            "description": "Section not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new answer in a section",
        "tags": [
          "Answers"
        ]
      }
    },
    "/api/v1/answers/{id}": {
      "put": {
        "operationId": "AnswersController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Answer UUID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAnswerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Answer updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not owner or admin"
          },
          "404": {
            "description": "Answer not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update an answer",
        "tags": [
          "Answers"
        ]
      },
      "delete": {
        "operationId": "AnswersController_delete_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Answer UUID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Answer deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteAnswerResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not owner or admin"
          },
          "404": {
            "description": "Answer not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete an answer",
        "tags": [
          "Answers"
        ]
      }
    },
...
  },
  "info": {
    "title": "SurveyScriber API",
    "description": "SurveyScriber Backend API - Professional Field Survey & Inspection Platform\n\n## Authentication\nThis API uses JWT Bearer tokens. Obtain tokens via `/auth/login` and include in the Authorization header.\n\n## Offline-First Architecture\nThe API supports granular CRUD operations for surveys, sections, and answers to enable offline-first Flutter clients to sync individual changes without full survey uploads.",
    "version": "1.0.0",
    "contact": {
      "name": "SurveyScriber Team",
      "url": "https://surveyscriber.com",
      "email": "support@surveyscriber.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token from /auth/login response",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateAnswerDto": {
        "type": "object",
        "properties": {
          "questionKey": {
            "type": "string",
            "example": "roof_condition",
            "description": "Unique key identifying the question"
          },
          "value": {
            "type": "string",
            "example": "Good condition with minor wear",
            "description": "The answer value"
          }
        },
        "required": [
          "questionKey",
          "value"
        ]
      },
...
      "AnswerResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "sectionId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "questionKey": {
            "type": "string",
            "example": "roof_condition"
          },
          "value": {
            "type": "string",
            "example": "Good condition with minor wear"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-15T10:30:00.000Z"
          }
        },
        "required": [
          "id",
          "sectionId",
          "questionKey",
          "value",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpdateAnswerDto": {
        "type": "object",
        "properties": {
          "questionKey": {
            "type": "string",
            "example": "roof_condition_updated",
            "description": "Unique key identifying the question"
          },
          "value": {
            "type": "string",
            "example": "Updated: Good condition with minor wear",
            "description": "The answer value"
          }
        }
      },
...
      "DeleteAnswerResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        },
        "required": [
          "success",
          "id"
        ]
      },
...
    }
  }
}